+++ packages/cli/vitest.config.ts
@@ -1,0 +1,18 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  resolve: {
+    alias: [
+      {
+        find: /^(\.{1,2}\/.*)\.js$/,
+        replacement: '$1',
+      },
+    ],
+  },
+  test: {
+    environment: 'node',
+    globals: true,
+    include: ['src/**/*.test.ts'],
+    setupFiles: ['./test-setup.ts'],
+  },
+});